Rename StatementGeneratorService class to StatementGeneratorHelper
